refactor: rename getPropError, change url in error
