+++ src/utils/getPropError/getPropError.ts
@@ -1,0 +1,5 @@
+const getPropError = (prop: string) => new ReferenceError(
+  `[is-scrolled]: ${prop} wasn't set properly. Take a look at the docs 🙌 https://github.com/bibixx/is-scrolled/#function-attributes`,
+);
+
+export default getPropError;
